backend/requirements.txt
# Django core
Django==4.2.27
djangorestframework==3.16.1
django-cors-headers==4.9.0
django-filter==25.1
django-tenants==3.9.0
drf-spectacular==0.29.0

# Database
psycopg2-binary==2.9.11

# Task Queue (database-based, no Redis needed)
django-q2==1.7.5

# AI
anthropic==0.76.0

# Server
gunicorn==23.0.0
whitenoise==6.11.0

# Utils
python-decouple==3.8
python-dotenv==1.2.1
pillow==12.1.0

# Required dependencies
anyio==4.12.1
asgiref==3.11.0
attrs==25.4.0
certifi==2026.1.4
click==8.3.1
colorama==0.4.6
distro==1.9.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
inflection==0.5.1
jiter==0.12.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
packaging==25.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.5
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
wcwidth==0.2.14

# For testing
requests==2.32.3
